working comp code & start of presentation

+++ Code/Test/signals/pwmwave.py
@@ -76,4 +76,4 @@ class PWMWave(Signal):
         self.duty_cycle = duty_cycle
 
     def evaluate(self, x: float):
-        return self.amplitude*(x % self.period < self.period * self.duty_cycle)
+        return (float)(self.amplitude*(x % self.period < self.period * self.duty_cycle))
